dependent: :destroy for exam uploads
This ensures that when an exam is destroyed, its upload is destroyed as well.  This does not prevent uploads from being destroyed without destroying the relevant exam, which can leave an exam with a nil upload. This will be fixed once exams handle their own files, instead of relying on another model.