Fix weird focus/ordering/side-effect issue where editing the message of a comment first, then editing the points of a comment, and then clicking on the comment, could cause the changed points to be lost
The callbacks all closed over stale variables, so updating all the `setCommentMap` calls to take in the current argument helped, but didn't completely solve the problem.  The needed fix was to
delay the creation of the new comment until all the other callbacks had been resolved, so that all the data were updated first.

(Credit: Jack Rosen for finding this issue)