Implementation of per-exam-version start/end/duration, fixes #416.
Technically, we could support the exam version only specifying a custom start or end or duration, rather than all three.  But for validation and for UI consistency, it's easier to specify all three at once.  Since we never refer to an exam's start/end/duration directly, but rather only through a registration's accommodated_start/end/duration, we simply interpose in those methods and refer to the exam_version's "effective" start/end/duration, which fall back to the exam's start/end/duration if they're not explicitly specified.