Fixes #442
New version ensures that:
* A question is extra credit if marked as such, OR if it is a single-part question and the sole part is extra-credit
* A part is extra credit if marked as such, OR if it is part of an extra-credit question.

I.e. if you author a question that is marked as extra credit, with two subparts A (not extra credit) and B (extra credit), then both parts will be counted as extra credit.  If you author a question that is not marked as extra credit, then individual parts within it can be marked as extra credit.