Add the ability to include code snippets, along with html snippets, as body items.
Some questions just need more than one code-snippet embedded in the prose...  This enables a single snippet at a time; it does not enable a multi-file display the way Exam, Question and Part each can show a multi-file FileViewer.  As with Code answers, the snippet can be edited directly or can be one of the uploaded files.